@@ -10,7 +10,7 @@ description: flutter_go
 version: 1.0.6
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -21,15 +21,16 @@ dependencies:
   cupertino_icons: ^0.1.2
   event_bus: ^1.0.1
   fluro: ^1.3.4
-  image_picker: ^0.6.0+1
+  image_picker: ^0.5.0
   sqflite: ^1.1.5
-  flutter_markdown: ^0.2.0
   url_launcher: ^5.0.2
   # 本地存储、收藏功能
   shared_preferences: ^0.4.3
+  share: ^0.6.1+1
   flutter_spinkit: "^3.1.0"
+  fluttertoast: ^3.1.0
   dio: ^2.0.15
-  flutter_webview_plugin: ^0.3.4
+  flutter_webview_plugin: ^0.3.5
   cookie_jar: ^1.0.0
   # 日期格式化
   intl: 0.15.7
@@ -40,11 +41,19 @@ dependencies:
   flutter_bloc: ^0.11.1
   bloc: ^0.12.0
   html: ^0.14.0+2
+  markdown: ^2.0.0
+  meta: ^1.0.5
+  string_scanner: ^1.0.0
+  path: ^1.5.1
   flutter_downloader: ^1.1.7
   path_provider: ^1.1.0
   permission_handler: ^3.0.0
   open_file: ^2.0.1+2
   package_info: ^0.4.0+3
+  flutter_jpush: ^0.0.4
+  zefyr:
+    path: ./zefyr
+
 
 dev_dependencies:
   flutter_test:
@@ -52,7 +61,6 @@ dev_dependencies:
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
 
@@ -203,6 +211,8 @@ flutter:
   - lib/widgets/themes/Cupertino/CupertinoTabScaffold/demo.dart
   - lib/widgets/themes/Cupertino/CupertinoTabView/demo.dart
   - lib/widgets/themes/Cupertino/CupertinoTimerPicker/demo.dart
+  - lib/page_demo_package/.demo.json
+  - lib/standard_pages/.pages.json
   - assets/app.db
   - assets/images/
   - assets/fonts/
