@@ -3,16 +3,17 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import '../routers/application.dart';
-import '../model/cat.dart';
+/// import '../model/cat.dart';
 import '../model/widget.dart';
 import '../widgets/index.dart';
 import '../components/widget_item_container.dart';
 
-enum CateOrWigdet { Cat, WidgetDemo }
+
 
 class CategoryHome extends StatefulWidget {
-  CategoryHome(this.name);
-  final String name;
+  CategoryHome(this.token);
+  final String token;
+
 
   @override
   _CategoryHome createState() => new _CategoryHome();
@@ -21,12 +22,11 @@ class CategoryHome extends StatefulWidget {
 class _CategoryHome extends State<CategoryHome> {
   String title = '';
   // 显示列表 cat or widget;
-  List<Cat> categories = [];
-  List<WidgetPoint> widgetPoints = [];
-  List<Cat> catHistory = new List();
+  List<CommonItem> items = [];
+  List<Object> widgetPoints = [];
+  List<CommonItem> catHistory = new List();
 
-  CatControlModel catControl = new CatControlModel();
-  WidgetControlModel widgetControl = new WidgetControlModel();
+
   // 所有的可用demos;
   List widgetDemosList = new WidgetDemoList().getDemos();
 
@@ -34,80 +34,60 @@ class _CategoryHome extends State<CategoryHome> {
   void initState() {
     super.initState();
     // 初始化加入顶级的name
-    this.getCatByName(widget.name).then((Cat cat) {
-      catHistory.add(cat);
-      searchCatOrWigdet();
+    print("这是新界面的id:>>> ${widget.token}");
+
+    CommonItem targetGroup = Application.widgetTree.find(widget.token) ?? [];
+    print("targetGroup::: $targetGroup");
+
+    catHistory.add(
+      targetGroup
+    );
+    this.setState(() {
+      items = targetGroup.children;
     });
+    searchCatOrWidget();
   }
 
-  Future<Cat> getCatByName(String name) async {
-    return await catControl.getCatByName(name);
-  }
+
 
   Future<bool> back() {
-    if (catHistory.length == 1) {
-      return Future<bool>.value(true);
-    }
-    catHistory.removeLast();
-    searchCatOrWigdet();
-    return Future<bool>.value(false);
+//    if (catHistory.length == 1) {
+//      return Future<bool>.value(true);
+//    }
+//    catHistory.removeLast();
+//    searchCatOrWidget();
+    return Future<bool>.value(true);
   }
 
-  void go(Cat cat) {
+  void go(CommonItem cat) {
     catHistory.add(cat);
-    searchCatOrWigdet();
+    searchCatOrWidget();
   }
 
-  void searchCatOrWigdet() async {
-    // 假设进入这个界面的parent一定存在
-    Cat parentCat = catHistory.last;
+  void searchCatOrWidget() async {
+    /// CommonItem widgetTree = Application.widgetTree;
+     // 假设进入这个界面的parent一定存在
+    CommonItem targetGroup = catHistory.last;
 
-
-    // 继续搜索显示下一级depth: depth + 1, parentId: parentCat.id
-    List<Cat> _categories =
-        await catControl.getList(new Cat(parentId: parentCat.id));
-    List<WidgetPoint> _widgetPoints = new List();
-    if (_categories.isEmpty) {
-      _widgetPoints =
-          await widgetControl.getList(new WidgetPoint(catId: parentCat.id));
-    }
 
     this.setState(() {
-      categories = _categories;
-      title = parentCat.name;
-      widgetPoints = _widgetPoints;
+      title = targetGroup.name;
     });
   }
 
-  void onCatgoryTap(Cat cat) {
+  void onCatgoryTap(CommonItem cat) {
     go(cat);
   }
 
-  void onWidgetTap(WidgetPoint widgetPoint) {
-    String targetName = widgetPoint.name;
-    String targetRouter = '/category/error/404';
-    widgetDemosList.forEach((item) {
-      // print("targetRouter = item.routerName> ${[item.name,targetName]}");
-      if (item.name == targetName) {
-        targetRouter = item.routerName;
-      }
-    });
-    Application.router.navigateTo(context, "$targetRouter");
-  }
+
 
   Widget _buildContent() {
     WidgetItemContainer wiContaienr = WidgetItemContainer(
       columnCount: 3,
-      categories: categories,
-      isWidgetPoint:false
+      commonItems: items
     );
-    if (widgetPoints.length > 0) {
-      wiContaienr = WidgetItemContainer(
-        categories: widgetPoints,
-        columnCount: 3,
-        isWidgetPoint:true
-      );
-    }
+
+
     return Container(
       padding: const EdgeInsets.only(bottom: 10.0, top: 5.0),
       decoration: BoxDecoration(
@@ -122,14 +102,18 @@ class _CategoryHome extends State<CategoryHome> {
 
   @override
   Widget build(BuildContext context) {
+
+
     return Scaffold(
       appBar: AppBar(
-        title: Text(title),
+        title: Text("$title"),
       ),
       body: WillPopScope(
+
         onWillPop: () {
           return back();
         },
+
         child: ListView(
           children: <Widget>[
             _buildContent(),
