@@ -4,11 +4,12 @@
 import 'dart:core';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_go/utils/data_utils.dart';
 
 import '../routers/application.dart';
 import '../routers/routers.dart';
 import '../components/markdown.dart';
-import '../model/collection.dart';
+/// import '../model/collection.dart';
 import '../widgets/index.dart';
 import '../event/event_bus.dart';
 import '../event/event_model.dart';
@@ -24,8 +25,8 @@ class WidgetDemo extends StatefulWidget {
       {Key key,
       @required this.title,
       @required this.contentList,
-      @required this.codeUrl,
-      @required this.docUrl,
+      this.codeUrl,
+      this.docUrl,
       this.bottomNaviBar})
       : super(key: key);
 
@@ -34,10 +35,10 @@ class WidgetDemo extends StatefulWidget {
 
 class _WidgetDemoState extends State<WidgetDemo> {
   bool _hasCollected = false;
-  CollectionControlModel _collectionControl = new CollectionControlModel();
+  /// CollectionControlModel _collectionControl = new CollectionControlModel();
   var _collectionIcons;
   List widgetDemosList = new WidgetDemoList().getDemos();
-  String _router = '';
+  String widgetType = 'old';
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
 
   List<Widget> _buildContent() {
@@ -64,56 +65,67 @@ class _WidgetDemoState extends State<WidgetDemo> {
   @override
   void initState() {
     super.initState();
-    _collectionControl.getRouterByName(widget.title).then((list) {
-      widgetDemosList.forEach((item) {
-        if (item.name == widget.title) {
-          _router = item.routerName;
+    // 这里不能直接 使用 ` ModalRoute.of(context)` 会产生报错
+    Future.delayed(Duration.zero, () {
+      String currentPath = ModalRoute.of(context).settings.name;
+      if (currentPath.indexOf('/standard-page') == 0) {
+        widgetType = 'standard';
+      }
+      Map<String, String> params = {
+        'type': widgetType,
+        "url": currentPath,
+        "name": widget.title
+      };
+      DataUtils.checkCollected(params).then((result) {
+        if (this.mounted) {
+          setState(() {
+            _hasCollected = result ?? null;
+          });
         }
       });
-      if (this.mounted) {
-        setState(() {
-          _hasCollected = list.length > 0;
-        });
-      }
     });
   }
 
 // 点击收藏按钮
   _getCollection() {
+    String currentRouterPath = ModalRoute.of(context).settings.name;
+    Map<String, String> params = {
+      "type": widgetType,
+      "url": currentRouterPath,
+      "name": widget.title
+    };
     if (_hasCollected) {
       // 删除操作
-      _collectionControl.deleteByName(widget.title).then((result) {
-        if (result > 0 && this.mounted) {
-          setState(() {
-            _hasCollected = false;
-          });
+      DataUtils.removeCollected(params, context).then((result) {
+        if (result) {
           _scaffoldKey.currentState
               .showSnackBar(SnackBar(content: Text('已取消收藏')));
           if (ApplicationEvent.event != null) {
             ApplicationEvent.event
-                .fire(CollectionEvent(widget.title, _router, true));
+                .fire(CollectionEvent(widget.title, currentRouterPath, true));
           }
-          return;
+          if (this.mounted) {
+            setState(() {
+              _hasCollected = false;
+            });
+          }
         }
-        print('删除错误');
       });
     } else {
       // 插入操作
-      _collectionControl
-          .insert(Collection(name: widget.title, router: _router))
-          .then((result) {  
-        if (this.mounted) {
-          setState(() {
-            _hasCollected = true;
-          });
-
-          if (ApplicationEvent.event != null) {
-            ApplicationEvent.event
-                .fire(CollectionEvent(widget.title, _router, false));
+      DataUtils.addCollected(params, context).then((result) {
+        if (result) {
+          if (this.mounted) {
+            setState(() {
+              _hasCollected = true;
+            });
           }
-
           _scaffoldKey.currentState
               .showSnackBar(SnackBar(content: Text('收藏成功')));
+          if (ApplicationEvent.event != null) {
+            ApplicationEvent.event
+                .fire(CollectionEvent(widget.title, currentRouterPath, false));
+          }
         }
       });
     }
@@ -129,7 +141,38 @@ class _WidgetDemoState extends State<WidgetDemo> {
           '${Routes.codeView}?filePath=${Uri.encodeComponent(widget.codeUrl)}');
     }
   }
-
+  List<PopupMenuEntry<String>> buildPopupMenu() {
+    List<PopupMenuEntry<String>> comps = [];
+    if (widget.docUrl != null) {
+      comps.add(
+          PopupMenuItem<String>(
+            value: 'doc',
+            child: ListTile(
+              leading: Icon(
+                Icons.library_books,
+                size: 22.0,
+              ),
+              title: Text('查看文档'),
+            ),
+          )
+      );
+    }
+    if (widget.codeUrl != null) {
+      comps.add(
+          PopupMenuItem<String>(
+            value: 'code',
+            child: ListTile(
+              leading: Icon(
+                Icons.code,
+                size: 22.0,
+              ),
+              title: Text('查看Demo'),
+            ),
+          )
+      );
+    }
+    return comps;
+  }
   @override
   Widget build(BuildContext context) {
     if (_hasCollected) {
@@ -137,50 +180,34 @@ class _WidgetDemoState extends State<WidgetDemo> {
     } else {
       _collectionIcons = Icons.favorite_border;
     }
+    List<PopupMenuEntry<String>> menus = buildPopupMenu();
+    List<Widget> actions = [
+      new IconButton(
+        tooltip: 'goBack home',
+        onPressed: () {
+          Navigator.popUntil(context, ModalRoute.withName('/'));
+        },
+        icon: Icon(Icons.home),
+      ),
+      new IconButton(
+        tooltip: 'collection',
+        onPressed: _getCollection,
+        icon: Icon(_collectionIcons),
+      ),
+    ];
+    if (menus.length > 0) {
+      actions.add(
+          PopupMenuButton<String>(
+            onSelected: _selectValue,
+            itemBuilder: (BuildContext context) => menus,
+          )
+      );
+    }
     return Scaffold(
         key: _scaffoldKey,
         appBar: AppBar(
           title: Text(widget.title),
-          actions: <Widget>[
-            new IconButton(
-              tooltip: 'goBack home',
-              onPressed: () {
-                Navigator.popUntil(context, ModalRoute.withName('/'));
-              },
-              icon: Icon(Icons.home),
-            ),
-            new IconButton(
-              tooltip: 'collection',
-              onPressed: _getCollection,
-              icon: Icon(_collectionIcons),
-            ),
-            PopupMenuButton<String>(
-              onSelected: _selectValue,
-              itemBuilder: (BuildContext context) => <PopupMenuEntry<String>>[
-                    const PopupMenuItem<String>(
-                      value: 'doc',
-                      child: ListTile(
-                        leading: Icon(
-                          Icons.library_books,
-                          size: 22.0,
-                        ),
-                        title: Text('查看文档'),
-                      ),
-                    ),
-                    const PopupMenuDivider(),
-                    const PopupMenuItem<String>(
-                      value: 'code',
-                      child: ListTile(
-                        leading: Icon(
-                          Icons.code,
-                          size: 22.0,
-                        ),
-                        title: Text('查看Demo'),
-                      ),
-                    ),
-                  ],
-            ),
-          ],
+          actions: actions,
         ),
         body: Container(
           padding: const EdgeInsets.symmetric(vertical: 10.0, horizontal: 15.0),
