@@ -6,24 +6,30 @@
 /// target:  app首页
 
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
+/// import 'package:flutter/rendering.dart';
+import 'package:flutter_go/utils/data_utils.dart';
 import 'package:flutter_go/utils/shared_preferences.dart';
-import 'package:flutter_go/views/first_page/first_page.dart';
+/// import 'package:flutter_go/views/first_page/first_page.dart';
 import 'package:flutter_go/views/first_page/main_page.dart';
+import 'package:fluro/fluro.dart';
+import 'package:flutter_go/views/user_page/user_page.dart';
 import 'package:flutter_go/views/widget_page/widget_page.dart';
 import 'package:flutter_go/views/welcome_page/fourth_page.dart';
-import 'package:flutter_go/views/collection_page/collection_page.dart';
+/// import 'package:flutter_go/views/collection_page/collection_page.dart';
 import 'package:flutter_go/routers/application.dart';
-import 'package:flutter_go/utils/provider.dart';
+/// import 'package:flutter_go/utils/provider.dart';
 import 'package:flutter_go/model/widget.dart';
-import 'package:flutter_go/widgets/index.dart';
+/// import 'package:flutter_go/widgets/index.dart';
 import 'package:flutter_go/components/search_input.dart';
 import 'package:flutter_go/model/search_history.dart';
 import 'package:flutter_go/resources/widget_name_to_icon.dart';
-
-const int ThemeColor = 0xFFC91B3A;
+import 'package:flutter_go/model/user_info.dart';
 
 class AppPage extends StatefulWidget {
+  final UserInformation userInfo;
+
+  AppPage(this.userInfo);
+
   @override
   State<StatefulWidget> createState() {
     return _MyHomePageState();
@@ -36,22 +42,32 @@ class _MyHomePageState extends State<AppPage>
   WidgetControlModel widgetControl = new WidgetControlModel();
   SearchHistoryList searchHistoryList;
   bool isSearch = false;
-  String appBarTitle = tabData[0]['text'];
+
   List<Widget> _list = List();
   int _currentIndex = 0;
-  static List tabData = [
-//    {'text': '业界动态', 'icon': Icon(Icons.language)},
+  List tabData = [
     {'text': 'WIDGET', 'icon': Icon(Icons.extension)},
-    {'text': '组件收藏', 'icon': Icon(Icons.favorite)},
     {'text': '关于手册', 'icon': Icon(Icons.import_contacts)},
+    {'text': '个人中心', 'icon': Icon(Icons.account_circle)},
+    //https://flutter-go.pub/api/isInfoOpen
   ];
-
   List<BottomNavigationBarItem> _myTabs = [];
+  String appBarTitle;
 
   @override
   void initState() {
     super.initState();
+    print('widget.userInfo    ${widget.userInfo}');
     initSearchHistory();
+
+    if(Application.pageIsOpen == true){// 是否展开业界动态
+      tabData.insert(0, {'text': '业界动态', 'icon': Icon(Icons.language)});
+      _list
+      //..add(FirstPage())
+        ..add(MainPage(userInfo: widget.userInfo));
+    }
+    appBarTitle = tabData[0]['text'];
+
     for (int i = 0; i < tabData.length; i++) {
       _myTabs.add(BottomNavigationBarItem(
         icon: tabData[i]['icon'],
@@ -60,12 +76,10 @@ class _MyHomePageState extends State<AppPage>
         ),
       ));
     }
-    _list
-//      ..add(FirstPage())
-//      ..add(MainPage())
-      ..add(WidgetPage(Provider.db))
-      ..add(CollectionPage())
-      ..add(FourthPage());
+      _list
+        ..add(WidgetPage())
+        ..add(FourthPage())
+        ..add(UserPage(userInfo: widget.userInfo));
   }
 
   @override
@@ -81,26 +95,21 @@ class _MyHomePageState extends State<AppPage>
   }
 
   void onWidgetTap(WidgetPoint widgetPoint, BuildContext context) {
-    List widgetDemosList = new WidgetDemoList().getDemos();
     String targetName = widgetPoint.name;
-    String targetRouter = '/category/error/404';
-    widgetDemosList.forEach((item) {
-      if (item.name == targetName) {
-        targetRouter = item.routerName;
-      }
-    });
-    searchHistoryList
-        .add(SearchHistory(name: targetName, targetRouter: targetRouter));
+    searchHistoryList.add(
+        SearchHistory(name: targetName, targetRouter: widgetPoint.routerName));
     print("searchHistoryList1 ${searchHistoryList.toString()}");
-    print("searchHistoryList2 ${targetRouter}");
-    print("searchHistoryList3 ${widgetPoint.name}");
-    Application.router.navigateTo(context, "$targetRouter");
+    String targetRouter = widgetPoint.routerName;
+    Application.router.navigateTo(context, targetRouter.toLowerCase(),
+        transition: TransitionType.inFromRight);
   }
 
   Widget buildSearchInput(BuildContext context) {
     return new SearchInput((value) async {
       if (value != '') {
-        List<WidgetPoint> list = await widgetControl.search(value);
+        print('value ::: $value');
+        // List<WidgetPoint> list = await widgetControl.search(value);
+        List<WidgetPoint> list = await DataUtils.searchWidget(value);
         return list
             .map((item) => new MaterialSearchResult<String>(
                   value: item.name,
@@ -118,11 +127,11 @@ class _MyHomePageState extends State<AppPage>
   }
 
   renderAppBar(BuildContext context, Widget widget, int index) {
-//    print('renderAppBar=====>>>>>>${index}');
-//    if (index == 0) {
-//      return null;
-//    }
-    return AppBar(title: buildSearchInput(context));
+    if (index == 1 && Application.pageIsOpen == true) {
+      return AppBar(title: buildSearchInput(context));
+    }else if(index == 0 && Application.pageIsOpen == false){
+      return AppBar(title: buildSearchInput(context));
+    }
   }
 
   @override
@@ -143,7 +152,7 @@ class _MyHomePageState extends State<AppPage>
         //fixed：固定
         type: BottomNavigationBarType.fixed,
 
-        fixedColor: Color(0xFFC91B3A),
+        fixedColor: Theme.of(context).primaryColor,
       ),
     );
   }
