@@ -95,16 +95,16 @@ class _CupertinoAppFullDefault extends State {
           ],
           onGenerateRoute: (setting) {
             // 当通过Navigation.of(context).pushNamed跳转路由时，在routes查找不到时，会调用该方法
-            routerHandler(setting);
+            return routerHandler(setting);
           },
           onGenerateTitle: (context) {
             // 跟上面的tiitle一样，但含有一个context参数用于做本地化
             return 'Flutter应用';
           },
-          onUnknownRoute: (setting) {
-            // 效果跟onGenerateRoute一样调用顺序为onGenerateRoute ==> onUnknownRoute
-            //RouterHandler(setting);
-          },
+//          onUnknownRoute: (setting) {
+//            // 效果跟onGenerateRoute一样调用顺序为onGenerateRoute ==> onUnknownRoute
+//            //RouterHandler(setting);
+//          },
           routes: {
             // 声明程序中有哪个通过Navigation.of(context).pushNamed跳转的路由参数以键值对的形式传递key:路由名字value:对应的Widget
             '/home': (BuildContext context) => HomePage(),
